Implement the Batch Processing Poset Directed Acyclic Graph
To connect to the server via Unix sockets, host should be set to the absolute path of the directory containing the socket file. Since / is a reserved character in URLs, the path should be URL encoded. If the path contains non-UTF 8 characters, a ConnectParams struct should be created manually and passed in. Note that Postgres does not support TLS over Unix sockets.Executes a statement, returning the number of rows modified.

A statement may contain parameters, specified by $n where n is the index of the parameter in the list provided, 1-indexed.

If the statement does not modify any rows (e.g. SELECT), 0 is returned.

If the same statement will be repeatedly executed (perhaps with different query parameters), consider using the prepare and prepare_cached methods.