lets get rid of this mock up
we will still approach the latency issue for extremely fast QPS with a bit of memoization, but remember we're focused on a loosely-coupled, transitive, irreflexive, and asymmetrical sync-free non-blocking chunck processing with no concept of row or table. We will get rid of the rows and oclumns before the first alpha, I want you to start thinking about solitons as entities and values as smart pointers with labels; a directed graph traversal starting from the reverse.