Add Lmdb extensions
remmeber we're going to want to address the database as a borrowed contextual multiplexed LOCAL entity. This means the state of the database is treated as any 0bject would, usually as a generic isolated namespace instance of; thus said, the way rust compiles allows for lazy appends through the instruction set via borrowed reference counters and memory reclamation protocols. For now let's focus on the wrapper and how the bitmask indicates the orientation of the data at hand