Interpret DAG as streams.
A bitstream comprised of Transaction headers in the packet and packet shuffler indicate the set of tokens embedded in the stream, have a root and some abstract stop codon for a partial order or semi order conditioned imposed on the tupleset embedded in the byte signature of the callback response. In this case, said posetDag is inferred to be in-place, and of logarithmic tailedness. (more on this later)