SQlite EDN wrapper
Here's what the above class is doing:
1. The `from_sql_value_pair` method takes a `rusqlite::types::Value` and a `value_type_tag` and returns a `TypedValue`.
2. The `to_sql_value_pair` method returns a tuple of `(ToSqlOutput, value_type_tag)`.
3. The `from_edn_value` method takes an `edn::Value` and returns an `Option<TypedValue>`.
4. The `to_edn_value_pair` method returns a tuple of `(edn::Value, ValueType)`.

The `from_sql_value_pair` method is the most complex. It takes a `rusqlite::types::Value` and a `value_type_tag` and returns a `TypedValue`. The `value_type_tag` is an integer that indicates the type of the `rusqlite::types::Value`. The `rusqlite::types::Value` is a `rusqlite::types::Value` that is used to represent the value in SQLite. The `rusqlite::types::Value` is a union of a few different types, so we have to match on the `value_type_tag` and the `rusqlite::types::Value` to figure out what to do.

The `value_type_tag`s are:

- 0: `rusqlite::types::Value::Integer`
- 1: `rusqlite::types::Value::Integer`
- 4: `rusqlite::types::Value::Integer`
- 5: `rusqlite::types::Value::Integer` or `rusqlite::types::Value::Real`
- 10: `rusqlite::types::Value::Text`
- 11: `rusqlite::types::Value::Blob`
- 13: `rusqlite::types::Value::Text`