EDN
EDN is a rust implementation which acts as a transducer of Clojure's Grammar rules. At Mozilla it became the foundation of Mentat, but EinstienDB is expanding Mozilla's use with the aspiration of turning it into an ML language in 2022.

EDN allows the implementation FromValue<Variable> for Variable.It then uses the `FromValue` trait to convert from a ValueAndSpan into an Option<Variable>. If it's not a PlainSymbol, return None (because we don't know how to make one of that type). Otherwise, if the string begins with '?', then wrap it in a Variable and return Some(it). Otherwise, just return None because it's not valid as a variable name.