SQUUIDS!
If the schema has an :einsteindb/index attribute for `attr`, return it's value as an integer (or return None).Otherwise, return None and do not modify the schema. This is used in tests to find indexes that are marked with :einsteindb/unique [:einsteindb/unique :value]. For example, (:foo/bar {:einsteindb/index true}) would be returned as Some(0), but (:foo/bar {:db/unique [:db/unique :value]}) would be returned as None since there is no :einsteindb/index attribute on this attr.
Note that this method only considers attributes that are indexed by virtue of being in the `INDEXED` set -it does not consider other kinds of indexes such as those created with :einsteindb/index true or those created with dupsort=true -just indexes created with :einsteindb/unique [:einsteindb/unique ...].